validation is added

+++ controller/bookmark.js
@@ -3,9 +3,53 @@ const Bookmark = require("../model/bookmark");
 module.exports = {
     addBookmark: async (req, res) => {
         try {
-            const userId = await req.body.userId;
-            const { questionId } = req.body;
+            
+    if (Object.keys(req.body).length === 0) {
+        return res
+            .status(406)
+            .json({
+                status: 406,
+                message: "Data not Found, Payload Not Acceptable",
+            });
+    }
+            let userId = await req.body.userId;
+            let { questionId } = req.body;
+            if(userId === undefined){
+                return res
+                .status(406)
+                .json({
+                    status: 406,
+                    message: "userId is not defined",
+                });
+            }
+            userId = userId.trim()
+            if(userId.length === 0){
+                return res
+                .status(406)
+                .json({
+                    status: 406,
+                    message: "please enter valid user Id",
+                });
+            }
+            if(questionId === undefined){
+                return res
+                .status(406)
+                .json({
+                    status: 406,
+                    message: "please enter the Quesiton Id",
+                });
+            }
+            questionId  = questionId.trim()
+            if(questionId.length === 0){
+                return res
+                .status(406)
+                .json({
+                    status: 406,
+                    message: "QuestionId can't be empty",
+                });
+            }
             const addedBookmark = await Bookmark.findOne({ userId, questionId });
+
             if (addedBookmark) {
                 // eslint-disable-next-line no-underscore-dangle
                 await Bookmark.findByIdAndDelete(addedBookmark._id);
